Only start Planktoscope MQTT clients for enabled controllers
