+++ internal/clients/instruments/queries/insert-camera.sql
@@ -1,2 +1,2 @@
-insert into instruments_camera (url, protocol, instrument_id, enabled)
-values ($url, $protocol, $instrument_id, $enabled);
+insert into instruments_camera (instrument_id, enabled, protocol, url)
+values ($instrument_id, $enabled, $protocol, $url);
