+++ internal/clients/instruments/queries/insert-controller.sql
@@ -1,2 +1,2 @@
-insert into instruments_controller (url, protocol, instrument_id, enabled)
-values ($url, $protocol, $instrument_id, $enabled);
+insert into instruments_controller (instrument_id, enabled, protocol, url)
+values ($instrument_id, $enabled, $protocol, $url);
