+++ internal/clients/instruments/queries/select-camera.sql
@@ -1,9 +1,9 @@
 select
   id            as id,
   instrument_id as instrument_id,
-  url           as url,
+  enabled       as enabled,
   protocol      as protocol,
-  enabled       as enabled
+  url           as url
 from instruments_camera as c
 where
   c.id = $id
