+++ internal/clients/instruments/queries/select-instruments.sql
@@ -4,13 +4,13 @@ select
   i.description       as description,
   i.admin_identity_id as admin_id,
   ca.id               as camera_id,
-  ca.url              as camera_url,
-  ca.protocol         as camera_protocol,
   ca.enabled          as camera_enabled,
+  ca.protocol         as camera_protocol,
+  ca.url              as camera_url,
   co.id               as controller_id,
-  co.url              as controller_url,
+  co.enabled          as controller_enabled,
   co.protocol         as controller_protocol,
-  co.enabled          as controller_enabled
+  co.url              as controller_url
 from instruments_instrument as i
 left join instruments_camera as ca
   on i.id = ca.instrument_id
