+++ internal/clients/instruments/queries/update-controller.sql
@@ -1,6 +1,6 @@
 update instruments_controller
 set
-  url = $url,
+  enabled = $enabled,
   protocol = $protocol,
-  enabled = $enabled
+  url = $url
 where instruments_controller.id = $id
