+++ web/templates/instruments/config/camera.partial.tmpl
