+++ web/templates/instruments/config/controller.partial.tmpl
